ed25519: factor out `hex` module
Note: should probably be replaced with `base16ct` as an optional
dependency, but this is a quick cleanup.
